Update the % money supply held by top 100 accouts for XRP:
The data source continues to be https://ledger.exposed/rich-stats

The percentage is calculated as follows:

 sum(top 100 wallets)
-----------------------
      (total XRP)

The calculation is against the total amount of XRP available. This
amount will decrease over time as XRP used to pay for fees is
destroyed.

As of this writing, the numbers are:

Sum(top 100 wallets exclusive of escrow):   38,337,799,799 XRP
   Sum(all wallets, exclusive of escrow):   47,291,885,259 XRP
                     Total available XRP:   99,992,216,935 XRP